Fix type mismatch: cast ssl_err from u32 to u64
- On Windows, unsigned long is 32 bits, so bindgen maps ssl_err to u32
- Error struct expects Option<u64> for tls_error field
- Cast c_err.ssl_err to u64 when creating tls_error field
- Fixes compilation error on Windows CI

+++ bindings/rust/mtls/src/error.rs
@@ -360,7 +360,7 @@ impl Error {
         };
 
         let tls_error = if c_err.ssl_err != 0 {
-            Some(c_err.ssl_err)
+            Some(c_err.ssl_err as u64)
         } else {
             None
         };
